# api_docs/openapi_app0.3.0_api_v1.yaml
openapi: 3.1.0
info:
  title: Health Monitor Backend
  version: 0.1.0
paths:
  /api/v1/blood-pressure/:
    get:
      tags:
      - blood-pressure
      summary: List Blood Pressure
      operationId: list_blood_pressure_api_v1_blood_pressure__get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/BloodPressureRead'
                type: array
                title: Response List Blood Pressure Api V1 Blood Pressure  Get
    post:
      tags:
      - blood-pressure
      summary: Create Blood Pressure
      operationId: create_blood_pressure_api_v1_blood_pressure__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BloodPressureCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BloodPressureRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/blood-pressure/import:
    post:
      tags:
      - blood-pressure
      summary: Import Blood Pressure
      operationId: import_blood_pressure_api_v1_blood_pressure_import_post
      requestBody:
        content:
          application/json:
            schema:
              items:
                $ref: '#/components/schemas/BloodPressureCreate'
              type: array
              title: Data
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/BloodPressureRead'
                type: array
                title: Response Import Blood Pressure Api V1 Blood Pressure Import
                  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/blood-pressure/chart:
    get:
      tags:
      - blood-pressure
      summary: Blood Pressure Chart
      operationId: blood_pressure_chart_api_v1_blood_pressure_chart_get
      parameters:
      - name: start
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Start
      - name: end
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: End
      responses:
        '200':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/blood-pressure/{record_id}:
    get:
      tags:
      - blood-pressure
      summary: Get Blood Pressure
      operationId: get_blood_pressure_api_v1_blood_pressure__record_id__get
      parameters:
      - name: record_id
        in: path
        required: true
        schema:
          type: integer
          title: Record Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BloodPressureRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - blood-pressure
      summary: Update Blood Pressure
      operationId: update_blood_pressure_api_v1_blood_pressure__record_id__put
      parameters:
      - name: record_id
        in: path
        required: true
        schema:
          type: integer
          title: Record Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BloodPressureUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BloodPressureRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - blood-pressure
      summary: Delete Blood Pressure
      operationId: delete_blood_pressure_api_v1_blood_pressure__record_id__delete
      parameters:
      - name: record_id
        in: path
        required: true
        schema:
          type: integer
          title: Record Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/blood-glucose/:
    get:
      tags:
      - blood-glucose
      summary: List Blood Glucose
      operationId: list_blood_glucose_api_v1_blood_glucose__get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/BloodGlucoseRead'
                type: array
                title: Response List Blood Glucose Api V1 Blood Glucose  Get
    post:
      tags:
      - blood-glucose
      summary: Create Blood Glucose
      operationId: create_blood_glucose_api_v1_blood_glucose__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BloodGlucoseCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BloodGlucoseRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/blood-glucose/import:
    post:
      tags:
      - blood-glucose
      summary: Import Blood Glucose
      operationId: import_blood_glucose_api_v1_blood_glucose_import_post
      requestBody:
        content:
          application/json:
            schema:
              items:
                $ref: '#/components/schemas/BloodGlucoseCreate'
              type: array
              title: Data
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/BloodGlucoseRead'
                type: array
                title: Response Import Blood Glucose Api V1 Blood Glucose Import Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/blood-glucose/chart:
    get:
      tags:
      - blood-glucose
      summary: Blood Glucose Chart
      operationId: blood_glucose_chart_api_v1_blood_glucose_chart_get
      parameters:
      - name: start
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Start
      - name: end
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: End
      responses:
        '200':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/blood-glucose/{record_id}:
    get:
      tags:
      - blood-glucose
      summary: Get Blood Glucose
      operationId: get_blood_glucose_api_v1_blood_glucose__record_id__get
      parameters:
      - name: record_id
        in: path
        required: true
        schema:
          type: integer
          title: Record Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BloodGlucoseRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - blood-glucose
      summary: Update Blood Glucose
      operationId: update_blood_glucose_api_v1_blood_glucose__record_id__put
      parameters:
      - name: record_id
        in: path
        required: true
        schema:
          type: integer
          title: Record Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BloodGlucoseUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BloodGlucoseRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - blood-glucose
      summary: Delete Blood Glucose
      operationId: delete_blood_glucose_api_v1_blood_glucose__record_id__delete
      parameters:
      - name: record_id
        in: path
        required: true
        schema:
          type: integer
          title: Record Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/ketones/:
    get:
      tags:
      - ketones
      summary: List Ketones
      operationId: list_ketones_api_v1_ketones__get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/KetonesRead'
                type: array
                title: Response List Ketones Api V1 Ketones  Get
    post:
      tags:
      - ketones
      summary: Create Ketones
      operationId: create_ketones_api_v1_ketones__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KetonesCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KetonesRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/ketones/import:
    post:
      tags:
      - ketones
      summary: Import Ketones
      operationId: import_ketones_api_v1_ketones_import_post
      requestBody:
        content:
          application/json:
            schema:
              items:
                $ref: '#/components/schemas/KetonesCreate'
              type: array
              title: Data
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/KetonesRead'
                type: array
                title: Response Import Ketones Api V1 Ketones Import Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/ketones/chart:
    get:
      tags:
      - ketones
      summary: Ketones Chart
      operationId: ketones_chart_api_v1_ketones_chart_get
      parameters:
      - name: start
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Start
      - name: end
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: End
      responses:
        '200':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/ketones/{record_id}:
    get:
      tags:
      - ketones
      summary: Get Ketones
      operationId: get_ketones_api_v1_ketones__record_id__get
      parameters:
      - name: record_id
        in: path
        required: true
        schema:
          type: integer
          title: Record Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KetonesRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - ketones
      summary: Update Ketones
      operationId: update_ketones_api_v1_ketones__record_id__put
      parameters:
      - name: record_id
        in: path
        required: true
        schema:
          type: integer
          title: Record Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KetonesUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KetonesRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - ketones
      summary: Delete Ketones
      operationId: delete_ketones_api_v1_ketones__record_id__delete
      parameters:
      - name: record_id
        in: path
        required: true
        schema:
          type: integer
          title: Record Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    BloodGlucoseCreate:
      properties:
        value:
          anyOf:
          - type: number
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Value
        measured_at:
          type: string
          format: date-time
          title: Measured At
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
      type: object
      required:
      - value
      - measured_at
      title: BloodGlucoseCreate
    BloodGlucoseRead:
      properties:
        value:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Value
        measured_at:
          type: string
          format: date-time
          title: Measured At
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        id:
          type: integer
          title: Id
      type: object
      required:
      - value
      - measured_at
      - id
      title: BloodGlucoseRead
    BloodGlucoseUpdate:
      properties:
        value:
          anyOf:
          - type: number
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Value
        measured_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Measured At
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
      type: object
      title: BloodGlucoseUpdate
    BloodPressureCreate:
      properties:
        systolic:
          type: integer
          title: Systolic
        diastolic:
          type: integer
          title: Diastolic
        pulse:
          anyOf:
          - type: integer
          - type: 'null'
          title: Pulse
        measured_at:
          type: string
          format: date-time
          title: Measured At
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
      type: object
      required:
      - systolic
      - diastolic
      - measured_at
      title: BloodPressureCreate
    BloodPressureRead:
      properties:
        systolic:
          type: integer
          title: Systolic
        diastolic:
          type: integer
          title: Diastolic
        pulse:
          anyOf:
          - type: integer
          - type: 'null'
          title: Pulse
        measured_at:
          type: string
          format: date-time
          title: Measured At
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        id:
          type: integer
          title: Id
      type: object
      required:
      - systolic
      - diastolic
      - measured_at
      - id
      title: BloodPressureRead
    BloodPressureUpdate:
      properties:
        systolic:
          anyOf:
          - type: integer
          - type: 'null'
          title: Systolic
        diastolic:
          anyOf:
          - type: integer
          - type: 'null'
          title: Diastolic
        pulse:
          anyOf:
          - type: integer
          - type: 'null'
          title: Pulse
        measured_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Measured At
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
      type: object
      title: BloodPressureUpdate
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    KetonesCreate:
      properties:
        value:
          anyOf:
          - type: number
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Value
        measured_at:
          type: string
          format: date-time
          title: Measured At
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
      type: object
      required:
      - value
      - measured_at
      title: KetonesCreate
    KetonesRead:
      properties:
        value:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Value
        measured_at:
          type: string
          format: date-time
          title: Measured At
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        id:
          type: integer
          title: Id
      type: object
      required:
      - value
      - measured_at
      - id
      title: KetonesRead
    KetonesUpdate:
      properties:
        value:
          anyOf:
          - type: number
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Value
        measured_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Measured At
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
      type: object
      title: KetonesUpdate
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError